kbs/resource: Add splitapi plugin
This plugin (based on the Nebula plugin interface) delivers credentials (keys
and certificates) to a sandbox (i.e., confidential PODs or VMs), specifically
to the kata agent to initiate the SplitAPI proxy server so that a workload
owner can communicate with the proxy server using a secure tunnel.

The proxy server's credentials can be requested using the kbs-client:

kbs-client --url http://127.0.0.1:8080 \
           get-resource \
           --path 'plugin/splitapi/credential?ip=60.11.12.13&name=pod6&id=32348'

The IPv4 address, name, and the ID of the sandbox must be provided in the query
string to obtain the credential resources from the kbs.

After receiving the credential request, the splitapi plugin will create a key
pair for the server and client and sign them using the self-signed CA. The
generated ca.crt, server.crt, and server.key are stored in a directory specific
to the sandbox (the caller) and returned to the caller. In addition, ca.key,
client.key, and client.crt are also generated and stored to that particular
directory specific to the sandbox (i.e., caller)

During the credential generation, a directory manager creates a unique
directory specific to the sandbox (i.e., the caller). The manager creates the
unique directory using the sandbox parameters passed in the query string.
A mapping file is also maintained to store the mapping between the sandbox
name and the unique directory created for the sandbox.

The splitapi plugin itself is not initialized by default. To initialize it,
you need to add 'splitapi' to 'manager_plugin_config.enabled_plugins' in
the kbs-config.toml.

Depends on: #451

Signed-off-by: Salman Ahmed <sahmed@ibm.com>